remove eigen

+++ src/lib/common.h
@@ -6,7 +6,7 @@
 
 namespace tlz {
 
-using real = float; // TODO double (change point)
+using real = double;
 using byte = std::uint8_t;
 using uchar = std::uint8_t;
 using ushort = std::uint16_t;
